requirements.txt
torch==2.8.0
torchvision
torchmetrics
transformers==4.56.1
jaxtyping
pandas
scikit-learn
scikit-image
opencv-python
matplotlib
open-clip-torch
pillow
beartype
h5py
tensorboard
dacite
peft
torcheval
typer
termcolor